@@ -3,51 +3,47 @@
 	"version": "0.0.1",
 	"private": true,
 	"scripts": {
-		"dev": "vite dev",
-		"build": "vite build",
-		"preview": "vite preview",
+		"dev": "bunx --bun vite dev",
+		"build": "bunx --bun vite build",
+		"preview": "bunx --bun vite preview",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@skeletonlabs/skeleton": "2.10.4",
-		"@skeletonlabs/tw-plugin": "0.4.0",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/adapter-static": "^3.0.1",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/eslint": "^9.0.0",
-		"@typescript-eslint/eslint-plugin": "^8.0.0",
-		"@typescript-eslint/parser": "^8.0.0",
-		"autoprefixer": "10.4.20",
-		"eslint": "^9.0.0",
+		"@eslint/compat": "^1.2.5",
+		"@eslint/js": "^9.18.0",
+		"@sveltejs/adapter-static": "^3.0.8",
+		"@sveltejs/kit": "^2.15.2",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@tailwindcss/vite": "^4.0.0-beta.9",
+		"@types/bun": "^1.1.16",
+		"bits-ui": "^1.0.0-next.77",
+		"clsx": "^2.1.1",
+		"eslint": "^9.18.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"postcss": "8.4.49",
-		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.1.2",
-		"prettier-plugin-tailwindcss": "^0.6.0",
-		"svelte": "^4.2.19",
-		"svelte-check": "^4.0.0",
-		"svelte-headless-table": "^0.18.2",
-		"tailwind-scrollbar": "^3.1.0",
-		"tailwindcss": "3.4.17",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.1.7",
-		"vite-plugin-tailwind-purgecss": "0.3.5"
+		"eslint-plugin-svelte": "^2.46.1",
+		"globals": "^15.14.0",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.3",
+		"prettier-plugin-tailwindcss": "^0.6.9",
+		"svelte": "^5.17.3",
+		"svelte-check": "^4.1.4",
+		"tailwindcss": "^4.0.0-beta.9",
+		"tailwind-merge": "^2.6.0",
+		"tailwind-scrollbar": "4.0.0-beta.0",
+		"tailwind-variants": "^0.3.0",
+		"tailwindcss-animate": "^1.0.7",
+		"typescript": "^5.7.3",
+		"typescript-eslint": "^8.19.1",
+		"vite": "^6.0.7"
 	},
 	"type": "module",
 	"dependencies": {
-		"bits-ui": "^0.22.0",
-		"clsx": "^2.1.0",
-		"svelte-radix": "^1.1.0",
-		"tailwind-merge": "^2.2.1",
-		"tailwind-variants": "^0.3.0"
-	},
-	"optionalDependencies": {
-		"@rollup/rollup-linux-x64-gnu": "4.30.1"
+		"@tanstack/table-core": "^8.20.5",
+		"@wailsio/runtime": "3.0.0-alpha.29",
+		"lucide-svelte": "^0.471.0",
+		"svelte-radix": "^2.0.1"
 	}
 }
