@@ -1,29 +1,42 @@
 <script lang="ts">
-	import { cfg } from "$lib/store";
+	import { cfg } from "$lib/state.svelte";
 	import { Switch } from "$lib/components/ui/switch";
 	import * as Card from "$lib/components/ui/card";
 	import { Checkbox } from "$lib/components/ui/checkbox";
 	import { fade } from "svelte/transition";
 	import { Input } from "$lib/components/ui/input";
 	import { Label } from "$lib/components/ui/label";
+	import { TypedObject } from "$lib/utils.js";
+	import type { CommandsConfig } from "@/bindings/github.com/BridgeSenseDev/Dank-Memer-Grinder/config";
 
 	function formatString(input: string): string {
 		return input
 			.replace(/[0-9]{2,}/g, (match) => ` ${match} `)
 			.replace(/[^A-Z0-9][A-Z]/g, (match) => `${match[0]} ${match[1]}`)
 			.replace(/[A-Z][A-Z][^A-Z0-9]/g, (match) => `${match[0]} ${match[1]}${match[2]}`)
-			.replace(/ {2,}/g, (match) => " ")
+			.replace(/ {2,}/g, () => " ")
 			.replace(/\s./g, (match) => match.toUpperCase())
 			.replace(/^./, (match) => match.toUpperCase())
 			.trim();
 	}
 
-	function updateCfg(commandKey: string, optionKey: string, value: any) {
-		cfg.update((cfg) => {
-			cfg.commands[commandKey][optionKey] = value;
-			return cfg;
-		});
+	function updateCfg(
+		commandKey: string,
+		optionKey: string,
+		value: string[] | string | number | boolean | Event
+	) {
+		if (value instanceof Event) {
+			value = (value.target as HTMLInputElement).value;
+			if (value.includes(",")) {
+				value = value.split(",").map((val) => val.trim());
+			}
+		}
+
+		commands[commandKey][optionKey] = value;
 	}
+
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
+	const commands = cfg.c.commands as CommandsConfig & Record<string, any>;
 </script>
 
 <div
@@ -31,23 +44,23 @@
 	out:fade={{ duration: 100 }}
 	class="z-10 h-full space-y-4"
 >
-	{#if $cfg.gui}
-		{#each Object.keys($cfg.commands) as commandKey (commandKey)}
+	{#if cfg.c.gui}
+		{#each TypedObject.keys(commands) as commandKey (commandKey)}
 			<Card.Root
-				class={`${$cfg.commands[commandKey].state ? "border-primary bg-primary-foreground/40" : ""}`}
+				class={`${commands[commandKey].state ? "border-primary bg-primary-foreground/40" : ""}`}
 			>
 				<Card.Header>
 					<div class="flex items-center space-x-2">
-						<Card.Title tag="h1">{formatString(commandKey)}</Card.Title>
-						<Switch id={commandKey} bind:checked={$cfg.commands[commandKey].state} />
+						<Card.Title>{formatString(commandKey)}</Card.Title>
+						<Switch id={commandKey} bind:checked={commands[commandKey].state} />
 						<Label for={commandKey}>Enabled</Label>
 					</div>
 				</Card.Header>
 				<Card.Content>
 					<div class="flex flex-col space-y-2">
-						{#each Object.keys($cfg.commands[commandKey]) as optionKey (optionKey)}
+						{#each TypedObject.keys(commands[commandKey]) as optionKey (optionKey)}
 							{#if optionKey !== "state"}
-								{#if typeof $cfg.commands[commandKey][optionKey] === "string"}
+								{#if typeof commands[commandKey][optionKey] === "string"}
 									<div class="flex w-1/2 flex-row items-center space-x-2">
 										<Label class="whitespace-nowrap" for={`${commandKey}_${optionKey}`}>
 											{formatString(optionKey)}
@@ -55,12 +68,12 @@
 										<Input
 											type="text"
 											id={`${commandKey}_${optionKey}`}
-											value={$cfg.commands[commandKey][optionKey]}
-											on:input={(e) => updateCfg(commandKey, optionKey, e.target.value)}
+											value={commands[commandKey][optionKey]}
+											oninput={(e) => updateCfg(commandKey, optionKey, e)}
 										/>
 									</div>
 								{/if}
-								{#if typeof $cfg.commands[commandKey][optionKey] === "number"}
+								{#if typeof commands[commandKey][optionKey] === "number"}
 									<div class="flex w-1/2 flex-row items-center space-x-2">
 										<Label class="whitespace-nowrap" for={`${commandKey}_${optionKey}`}>
 											{formatString(optionKey)}
@@ -68,21 +81,24 @@
 										<Input
 											type="number"
 											id={`${commandKey}_${optionKey}`}
-											value={$cfg.commands[commandKey][optionKey]}
-											on:input={(e) => updateCfg(commandKey, optionKey, parseInt(e.target.value))}
+											value={commands[commandKey][optionKey]}
+											oninput={(e) => updateCfg(commandKey, optionKey, e)}
 										/>
 									</div>
-								{:else if typeof $cfg.commands[commandKey][optionKey] === "boolean"}
+								{:else if typeof commands[commandKey][optionKey] === "boolean"}
 									<div class="flex w-full flex-row items-center space-x-2">
 										<Checkbox
 											id={`${commandKey}_${optionKey}`}
-											bind:checked={$cfg.commands[commandKey][optionKey]}
+											bind:checked={commands[commandKey][optionKey]}
 										/>
-										<Label class="whitespace-nowrap" for={`${commandKey}_${optionKey}`}>
+										<Label
+											class="cursor-pointer whitespace-nowrap"
+											for={`${commandKey}_${optionKey}`}
+										>
 											{formatString(optionKey)}
 										</Label>
 									</div>
-								{:else if typeof $cfg.commands[commandKey][optionKey] === "object"}
+								{:else if typeof commands[commandKey][optionKey] === "object"}
 									<div class="flex w-full flex-row items-center space-x-2">
 										<Label class="whitespace-nowrap" for={`${commandKey}_${optionKey}`}>
 											{formatString(optionKey)}
@@ -91,25 +107,16 @@
 											<Input
 												type="text"
 												id={`${commandKey}_${optionKey}`}
-												value={$cfg.commands[commandKey][optionKey].join(", ")}
-												on:input={(e) => {
-													if (e.target.value) {
-														const values = e.target.value.split(",");
-														const numbers = values.map((val) =>
-															val ? (isNaN(Number(val)) ? 0 : Number(val)) : ""
-														);
-														updateCfg(commandKey, optionKey, numbers);
-													} else {
-														updateCfg(commandKey, optionKey, []);
-													}
+												value={commands[commandKey][optionKey].join(", ")}
+												oninput={(e) => {
+													updateCfg(commandKey, optionKey, e);
 												}}
 											/>
 										{:else}
 											<Input
-												value={$cfg.commands[commandKey][optionKey].join(", ")}
+												value={commands[commandKey][optionKey].join(", ")}
 												id={`${commandKey}_${optionKey}`}
-												on:input={(e) =>
-													updateCfg(commandKey, optionKey, e.target.value.split(", "))}
+												oninput={(e) => updateCfg(commandKey, optionKey, e)}
 											/>
 										{/if}
 									</div>
