@@ -1,15 +1,16 @@
 import { Dialog as DialogPrimitive } from "bits-ui";
 
-const Root = DialogPrimitive.Root;
-const Trigger = DialogPrimitive.Trigger;
-
 import Title from "./dialog-title.svelte";
-import Portal from "./dialog-portal.svelte";
 import Footer from "./dialog-footer.svelte";
 import Header from "./dialog-header.svelte";
 import Overlay from "./dialog-overlay.svelte";
 import Content from "./dialog-content.svelte";
 import Description from "./dialog-description.svelte";
+
+const Root: typeof DialogPrimitive.Root = DialogPrimitive.Root;
+const Trigger: typeof DialogPrimitive.Trigger = DialogPrimitive.Trigger;
+const Close: typeof DialogPrimitive.Close = DialogPrimitive.Close;
+const Portal: typeof DialogPrimitive.Portal = DialogPrimitive.Portal;
 
 export {
 	Root,
@@ -21,6 +22,7 @@ export {
 	Overlay,
 	Content,
 	Description,
+	Close,
 	//
 	Root as Dialog,
 	Title as DialogTitle,
@@ -30,5 +32,6 @@ export {
 	Trigger as DialogTrigger,
 	Overlay as DialogOverlay,
 	Content as DialogContent,
-	Description as DialogDescription
+	Description as DialogDescription,
+	Close as DialogClose
 };
