@@ -1,56 +1,18 @@
 <script lang="ts">
+	import { browser } from "$app/environment";
 	import * as Tabs from "$lib/components/ui/tabs/index.js";
 	import { fade } from "svelte/transition";
-	import { EventsOn } from "$lib/wailsjs/runtime/runtime";
-	import { onMount } from "svelte";
-	import { importantLogs, othersLogs, discordLogs, logsEventListenersSet } from "$lib/store";
+	import { logs } from "$lib/state.svelte";
 
-	let mainCurrentTab =
-		typeof window !== "undefined"
-			? sessionStorage.getItem("mainCurrentTab") || "important"
-			: "important";
+	let mainCurrentTab = $state("important");
 
-	{
-		if (typeof window !== "undefined") {
-			sessionStorage.setItem("mainCurrentTab", mainCurrentTab);
-		}
+	if (browser) {
+		mainCurrentTab = sessionStorage.getItem("mainCurrentTab") || "important";
 	}
 
-	onMount(async () => {
-		if (!$logsEventListenersSet) {
-			const logWithTime = (logs: string, type: string, username: string, msg: string) => {
-				const now = new Date();
-				const hours = now.getHours();
-				const minutes = now.getMinutes();
-				const ampm = hours >= 12 ? "PM" : "AM";
-				const formattedTime = `${hours % 12 || 12}:${minutes < 10 ? "0" : ""}${minutes}${ampm}`;
-
-				logs += `<span class="text-gray-400">${formattedTime} `;
-
-				if (type === "INF") {
-					logs += `<span class="text-green-500">INF `;
-				} else if (type === "ERR") {
-					logs += `<span class="text-red-500">ERR `;
-				}
-				logs +=
-					`<span class="text-sky-500">` + username + '<span class="text-white"> ' + msg + "<br>";
-
-				return logs;
-			};
-
-			EventsOn("logImportant", (type: string, username: string, msg: string) => {
-				importantLogs.update((logs) => logWithTime(logs, type, username, msg));
-			});
-
-			EventsOn("logOthers", (type: string, username: string, msg: string) => {
-				othersLogs.update((logs) => logWithTime(logs, type, username, msg));
-			});
-
-			EventsOn("logDiscord", (type: string, username: string, msg: string) => {
-				discordLogs.update((logs) => logWithTime(logs, type, username, msg));
-			});
-
-			logsEventListenersSet.set(true);
+	$effect(() => {
+		if (browser) {
+			sessionStorage.setItem("mainCurrentTab", mainCurrentTab);
 		}
 	});
 </script>
@@ -68,22 +30,22 @@
 		</Tabs.List>
 		<Tabs.Content value="important" class="h-full">
 			<div
-				class="flex h-full min-h-[60px] w-full rounded-md border border-input bg-transparent px-3 py-2 text-sm shadow-sm outline-none ring-1 ring-ring"
-				bind:innerHTML={$importantLogs}
+				class="border-input ring-ring flex h-full min-h-[60px] w-full rounded-md border bg-transparent px-3 py-2 text-sm ring-1 shadow-xs outline-hidden"
+				bind:innerHTML={logs.importantLogs}
 				contenteditable="false"
 			></div>
 		</Tabs.Content>
 		<Tabs.Content value="others" class="h-full">
 			<div
-				class="flex h-full min-h-[60px] w-full rounded-md border border-input bg-transparent px-3 py-2 text-sm shadow-sm outline-none ring-1 ring-ring"
-				bind:innerHTML={$othersLogs}
+				class="border-input ring-ring flex h-full min-h-[60px] w-full rounded-md border bg-transparent px-3 py-2 text-sm ring-1 shadow-xs outline-hidden"
+				bind:innerHTML={logs.othersLogs}
 				contenteditable="false"
 			></div>
 		</Tabs.Content>
 		<Tabs.Content value="discord" class="h-full">
 			<div
-				class="flex h-full min-h-[60px] w-full rounded-md border border-input bg-transparent px-3 py-2 text-sm shadow-sm outline-none ring-1 ring-ring"
-				bind:innerHTML={$discordLogs}
+				class="border-input ring-ring flex h-full min-h-[60px] w-full rounded-md border bg-transparent px-3 py-2 text-sm ring-1 shadow-xs outline-hidden"
+				bind:innerHTML={logs.discordLogs}
 				contenteditable="false"
 			></div>
 		</Tabs.Content>
