@@ -3,18 +3,24 @@ import { vitePreprocess } from "@sveltejs/vite-plugin-svelte";
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	extensions: [".svelte"],
-	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
-	// for more information about preprocessors
-	preprocess: [vitePreprocess()],
-
+	preprocess: vitePreprocess(),
 	kit: {
 		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
-		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
+		// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
 		adapter: adapter({
-			fallback: "index.html"
-		})
+			pages: "dist",
+			assets: "dist",
+			fallback: undefined,
+			precompress: false,
+			strict: true
+		}),
+		alias: {
+			$lib: "./src/lib",
+			"$lib/*": "./src/lib/*",
+			"@/*": "./*"
+		}
 	}
 };
+
 export default config;
